Better Pin handling
Our current code was taking pinned objects and just casting them to the
requisite types.  This works, but I think that it was by accident more
than design.  I believe that what was happening was that `Pin<Box<T>`,
`Box<T>`, and `*mut T` are all effectively the same value once all the
rust compiler decorations are erased. So a cast to a pointer from any of
these does work.  However, this is not something the API guarantees.
Though `Pin` and `Box` are both annotated with `#[repr(transparent)]`,
this is not a promise of a particular layout, but a hint.

The correct way to handle these cases is to use `Pin::as_mut()` to get a
`Pin<&mut T>` and then use `Pin::into_inner()` on that to get a mutable reference to the memory.

This pinning stuff is a little hard to grok still, but my understanding
is that using `Box` consistently, as we do, ensures that we don't have a
problem unless we try to move the object that the `Box` points at.
We get that as long as we only use these functions to hand out pointers
to the C code.