Rust 1.48 clippy changes (#1067)
* Rust 1.48 clippy changes

There was one here that was genuinely useful, but some of the others
were just annoying.  The insistence on using `Option::map_or` made
some of the code worse, so I suppressed it.  Also, qpack deliberately
overwrites specific error codes with `Result::map_err`, which now
generates warnings that needed to be suppressed.

* A whole bunch of problems that I somehow missed

* One error code mapping seemed wrong

This fixes the test for that.