Enable ECH (#1131)
* Enable ECH

There is a lot of plumbing here, but it's all pretty simple stuff.
Testing is mostly superficial.

I pulled in a bunch of code from my work on HPKE (as there are
commonalities there).  That means some improvements to neqo-crypto, in
particular, the types we use for wrapping NSS objects have been enhanced
a little.  Some of that is to support the use of the ECH API (PublicKey
and PrivateKey in particular), but I also improved our handling of
SECItem by taking some of the enhancements I made.

The only thing I didn't do was handle an automatic retry for the client
code when the ECH retry is used.  That's not very important here, but it
might be worth doing just so that it can be demonstrated.

* Fixups from self-review

* Fix ordering

* No need for extra alias

* Add a better API for fallbacks

This uses events for validating the public name.  This should make it
easier to integrate.

I've also fixed the alert codes to send ech_required as required by
the specification.

* Merge main back

* Merge main