Lots of lints (#1227)
* Lots of lints

It seems like a lot of code landed without any of these checks being
run.  Hopefully we can be stricter about this in future.  There were
some minor (but relevant) problems in there.

(Lots of missing semicolons, which I blame rustfmt for; and lots of
extra `&`, which I might have to own up to.)

* Address comments