Use a AES header protection context (#1307)
* Use a AES header protection context

This constructs a PK11Context for use with AES ciphers.  AES has a bunch
of initialization costs that using PK11_Encrypt does not amortize very
well.  This constructs the context once and reuses it for every header
mask/unmask operation.

This is a little fiddly in that Chacha20 doesn't have an interface that
would allow us to use it in the same way.  So that continues to use
PK11_Encrypt.

For https://bugzilla.mozilla.org/show_bug.cgi?id=1748097 with no real
knowledge (yet) about whether it helps or hinders.

* Cleanup