Cache random (#1640)
* Avoid allocation in random()

This makes the function take its argument as a const generic argument,
which allows the allocation to be performed on the stack.

There are a few cases where we need to do in-place randomization of a
different type of object (`SmallVec` in a few places) so I've also
exposed an in-place mutation function.

Next step is to cache the slot that this uses.

* Cache randomness

* Add a test for the cache

* Remove dead code

---------

Co-authored-by: Lars Eggert <lars@eggert.org>